inc/server can be empty

+++ phylogeny/distTree_inc_complete.sh
@@ -4,7 +4,7 @@ source $THIS/../bash_common.sh
 if [ $# -ne 2 ]; then
   echo "Compute a complete pair-wise dissimilarity matrix and build a distance tree using the incremental tree data structure"
   echo "#1: incremental distance tree directory"
-  echo "#2: list of objects"
+  echo "#2: sorted list of objects"
   echo "Output: #1/, #1/../data.dm"
   exit 1
 fi
@@ -25,30 +25,32 @@ if [ $N -gt 0 ]; then
   exit 1
 fi
 
+sort -c $OBJS
 
-$THIS/../sort.sh $OBJS
 
 section "Computing dissimilarities"
 $THIS/../list2pairs $OBJS > $INC/dissim_request
 $THIS/distTree_inc_request2dissim.sh $INC $INC/dissim_request $INC/dissim.raw
 rm $INC/dissim_request
-cat $INC/dissim.raw | grep -vwi "nan" | grep -vwi "inf" > $INC/dissim
+cat $INC/dissim.raw | grep -vwi "nan$" | grep -vwi "inf$" > $INC/dissim
 rm $INC/dissim.raw
 
 section "data.dm"
 $THIS/../dm/pairs2dm $INC/dissim 1 "dissim" 6 -distance > $INC/../data.dm
 echo "nan:"
 set +o errexit
-grep -wic 'nan' $INC/../data.dm
+grep -wic 'nan$' $INC/../data.dm
 set -o errexit
 
-section "Outliers"
 $THIS/../dm/dm2objs $INC/../data -noprogress | sort > $INC/tree.list
 
-$THIS/../setMinus $OBJS $INC/tree.list > $INC/outlier-alien
-wc -l $INC/outlier-alien
-$THIS/../trav $INC/outlier-alien "$INC/outlier2db.sh %f alien"
-rm $INC/outlier-alien
+if [ -s $INC/server ]; then
+  section "Outliers"
+  $THIS/../setMinus $OBJS $INC/tree.list > $INC/outlier-alien
+  wc -l $INC/outlier-alien
+  $THIS/../trav $INC/outlier-alien "$INC/outlier2db.sh %f alien"
+  rm $INC/outlier-alien
+fi
 
 
 HYBRIDNESS_MIN=`cat $INC/hybridness_min`
@@ -59,16 +61,14 @@ if [ $HYBRIDNESS_MIN != 0 ]; then
   $THIS/../dm/distTriangle $INC/../data "dissim"  -clustering_dir $INC/clust  -hybridness_min $HYBRIDNESS_MIN  -hybrid $INC/hybrid.new
   N=`ls $INC/clust/ | wc -l`
   if [ $N -gt 1 ]; then
-    echo "# Clusters: $N"
-    exit 1
+    error "# Clusters: $N"
   fi
   mv $INC/clust/1/data.dm $INC/../data.dm
   rm -r $INC/clust/
-  $THIS/../dm/dm2objs $INC/../data | sort > $INC/tree.list  
-  section "Hybrid"
-	$THIS/distTree_inc_hybrid.sh $INC 
- #echo "Unhybrid"
- #$THIS/distTree_inc_unhybrid.sh $INC 
+  if [ -s $INC/server ]; then
+    section "Hybrid"
+  	$THIS/distTree_inc_hybrid.sh $INC 
+  fi
 fi
 
 
@@ -81,13 +81,13 @@ fi
 VARIANCE=`cat $INC/variance`
 $THIS/makeDistTree  -threads 5  -data $INC/../data  -dissim_attr "dissim"  -variance $VARIANCE  -optimize  -subgraph_iter_max 10  $HYBRID  -output_tree $INC/tree  > $INC/hist/makeDistTree-complete.1
 
-section "Database"
-$INC/objects_in_tree.sh $INC/tree.list 1
-if [ $HYBRIDNESS_MIN != 0 ]; then
-  section "Hybrid"
-	$THIS/distTree_inc_hybrid.sh $INC
- #echo "Unhybrid"
- #$THIS/distTree_inc_unhybrid.sh $INC
+if [ -s $INC/server ]; then
+  section "Database"
+  $INC/objects_in_tree.sh $INC/tree.list 1
+  if [ $HYBRIDNESS_MIN != 0 ]; then
+    section "Hybrid"
+  	$THIS/distTree_inc_hybrid.sh $INC
+  fi
 fi
 
 rm $INC/tree.list
@@ -102,7 +102,7 @@ if [ -e $INC/phen ]; then
     LARGE=1
   fi
   $THIS/tree_quality_phen.sh $INC/tree "" $INC/phen $LARGE 0 "" > $INC/hist/tree_quality_phen.1
-	grep ' !' $INC/hist/tree_quality_phen.1
+	grep 'V !$' $INC/hist/tree_quality_phen.1
 fi
